{{cookiecutter.project_slug}}/requirements_dev.txt
#

# These requirements were autogenerated by pipenv

# To regenerate from the project's Pipfile, run:

#

#    pipenv lock --requirements --dev

#



# Note: in pipenv 2020.x, "--dev" changed to emit both default and development

# requirements. To emit only development requirements, pass "--dev-only".



-i https://pypi.org/simple

atomicwrites==1.4.0; sys_platform == 'win32'

attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

black==21.9b0

click==8.0.3; python_version >= '3.6'

colorama==0.4.4; sys_platform == 'win32' and platform_system == 'Windows'

flake8==4.0.1

iniconfig==1.1.1

isort==5.9.3

mccabe==0.6.1

mypy-extensions==0.4.3

mypy==0.910

packaging==21.0; python_version >= '3.6'

pathspec==0.9.0

platformdirs==2.4.0; python_version >= '3.6'

pluggy==1.0.0; python_version >= '3.6'

py==1.10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

pycodestyle==2.8.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

pyflakes==2.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'

pytest==6.2.5

regex==2021.10.8

toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'

tomli==1.2.1; python_version >= '3.6'

typing-extensions==3.10.0.2

pyyaml==5.4.1
